moving tunnel and server code to their own modules...added unit tests around them
