+++ test/core/cloud.test.js
@@ -201,63 +201,6 @@ describe('cloud', () => {
 
   it('should support sr', () => cloud.sr('/foo', (r) => expect(r).to.have.statusCode(200)));
 
-  it('should allow starting http server', () => {
-    const port = 8085;
-    let server;
-    return cloud.startServer(port)
-    .then(s => {
-      server = s;
-      chakram.setRequestDefaults({ headers: { 'User-Agent': 'churros-test' } });
-      return chakram.post('http://localhost:' + port, { event: 'green hat' })
-    })
-    .then(r => {
-      expect(r.response.statusCode).to.equal(502);
-    })
-    .then(r => server.destroy());
-  });
-
-  it('should support listening for events with custom validation', () => {
-    const port = 8085;
-    return cloud.startServer(port)
-    .then(server => {
-      return new Promise((res, rej) => {
-        cloud.listenForEvents(server, 1, 5)
-          .then(r => r.forEach(e => expect(e).to.not.be.empty))
-          .then(r => res(r))
-          .catch(r => rej('How embarrassing, I failed validating events: ' + r));
-        chakram.setRequestDefaults({ headers: { 'User-Agent': 'churros-test' } });
-        return chakram.post('http://localhost:' + port, { event: 'green hat' })
-      });
-    });
-  });
-
-  it('should throw an error if event validation fails', () => {
-    const port = 8085;
-    return cloud.startServer(port)
-    .then(server => {
-      return new Promise((res, rej) => {
-        cloud.listenForEvents(server, 1, 5)
-          .then(r => r.forEach(e => expect(e).to.be.empty))
-          .then(r => rej('How embarrassing, I should have failed validating events: ' + r))
-          .catch(r => res(r));
-        chakram.setRequestDefaults({ headers: { 'User-Agent': 'churros-test' } });
-        return chakram.post('http://localhost:' + port, { event: 'green hat' });
-      });
-    });
-  });
-
-  it('should throw an error if event listen times out', () => {
-    const port = 8085;
-    return cloud.startServer(port)
-    .then(server => {
-      return new Promise((res, rej) => {
-        cloud.listenForEvents(server, 1, 1)
-          .then(r => rej('How embarrassing, I should have timed out listening for events: ' + r))
-          .catch(r => res(r));
-      });
-    });
-  });
-
   it('should support post file', () => {
     // should really NOT depend on the file system here :/
     const filePath = '.tmp';
