+++ test/core/tools.test.js
@@ -51,13 +51,4 @@ describe('tools', () => {
     expect(encoded).to.be.a('string');
     expect(encoded).to.equal('ABCD');
   });
-
-  it('should support starting up a localtunnel', () => {
-    return tools.startTunnel(8999)
-      .then(tunnel => {
-        expect(tunnel).to.not.be.null;
-        expect(tunnel.url).to.be.a('string');
-        tunnel.close();
-      });
-  });
 });
